Move money calculation to a another class
