+++ src/Game/NeoServer.Game.Creatures/Player/Inventory/Calculations/InventoryAttackCalculation.cs
@@ -1,8 +1,6 @@
 ﻿using System;
-using NeoServer.Game.Common.Contracts.DataStores;
 using NeoServer.Game.Common.Contracts.Items.Types.Body;
 using NeoServer.Game.Common.Creatures.Players;
-using NeoServer.Game.Common.Item;
 
 namespace NeoServer.Game.Creatures.Player.Inventory.Calculations;
 
@@ -26,29 +24,7 @@ public static class InventoryCalculation
 
         return attack;
     }
-
-    internal static ulong CalculateTotalMoney(this Inventory inventory, ICoinTypeStore coinTypeStore)
-    {
-        uint total = 0;
-
-        var inventoryMap = inventory.BackpackSlot?.Map;
-        var coinTypes = coinTypeStore.All;
-
-        if (coinTypes is null) return total;
-        if (inventoryMap is null) return total;
-
-        foreach (var coinType in coinTypes)
-        {
-            if (coinType is null) continue;
-            if (!inventoryMap.TryGetValue(coinType.TypeId, out var coinAmount)) continue;
-
-            var worthMultiplier = coinType?.Attributes?.GetAttribute<uint>(ItemAttribute.Worth) ?? 0;
-            total += worthMultiplier * coinAmount;
-        }
-
-        return total;
-    }
-
+    
     internal static byte CalculateAttackRange(this InventoryMap inventoryMap)
     {
         var rangeLeft = 0;
